Fix compile

+++ courses/koopa_troopa_beach/course_offsets.c
@@ -130,7 +130,7 @@ Gfx* koopa_troopa_beach_dls2[] = {
 
 // 0x06014828
 // animations? lights?
-u32 d_course_koopa_troopa_beach_unk_data1[] = {
+uintptr_t d_course_koopa_troopa_beach_unk_data1[] = {
     0x00000000,
     0x00000001,
     0x00000000,
@@ -316,7 +316,7 @@ Animation* d_course_koopa_troopa_beach_unk_data5[] = {
 };
 
 // No idea what this stuff is
-u32 d_course_koopa_troopa_beach_unk_data6[] = {
+uintptr_t d_course_koopa_troopa_beach_unk_data6[] = {
     0x00000000, 0x00000000, 0x00000001, 0x00000000, 0x00000001,
 };
